apps/requirements.txt
asgiref==3.8.1
autopep8==2.3.2
dj-database-url==3.0.0
Django==5.2.3
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycodestyle==2.14.0
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.0
whitenoise==6.9.0
